requirements.txt
Flask
flask-cors
anthropic

# Production server
gunicorn

# Portal SSO
PyJWT[crypto]
cryptography
requests
